- updated Renew dialog text - renamed Payment dialog -> Confirm NR dialog
