+++ src/components/dialogs/index.ts
@@ -2,6 +2,7 @@ export { default as AdvancedSearch } from './advanced-search.vue'
 export { default as AffiliationErrorDialog } from './affiliation-error.vue'
 export { default as CancelDialog } from './cancel.vue'
 export { default as ConditionsDialog } from './conditions.vue'
+export { default as ConfirmNrDialog } from './confirm-name-request.vue'
 export { default as ErrorDialog } from './error.vue'
 export { default as ExitDialog } from './exit.vue'
 export { default as HelpMeChooseDialog } from './help-me-choose.vue'
@@ -9,7 +10,6 @@ export { default as LocationInfoDialog } from './location-info.vue'
 export { default as MrasSearchInfoDialog } from './mras-search-info.vue'
 export { default as NameCheckIssuesDialog } from './name-check-issues.vue'
 export { default as NrNotRequiredDialog } from './nr-not-required.vue'
-export { default as PaymentDialog } from './payment.vue'
 export { default as PaymentCompleteDialog } from './payment-complete.vue'
 export { default as PickEntityOrConversionDialog } from './pick-entity-or-conversion.vue'
 export { default as PickRequestTypeDialog } from './pick-request-type.vue'
